pyproject.toml
[project]
name = "retrieval-based-text-classification"
version = "0.0.1"
description = ""
authors = [
    {name = "Leroll"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.116.1",
    "fire>=0.7.0",
    "loguru>=0.7.3",
    "psutil>=7.0.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0.2",
    "uvicorn>=0.35.0",
]

[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple/"
package = true

[project.scripts]
rtc = "retrieval_based_text_classification.cli:cli"